+++ metagraph/experiments/lossless_dbg/dynamic_routing_table.hpp
@@ -39,7 +39,7 @@ class IdentityComparator : public Reference {
 public:
     using Reference::Reference;
     template<typename ...Args>
-    int rank(Args... args) const {
+    int64_t rank(Args... args) const {
         auto target = Reference::rank(args...);
         auto value = t.rank(args...);
         if (target != value) {
@@ -50,7 +50,7 @@ public:
     }
 
     template<typename ...Args>
-    int size(Args... args) const {
+    int64_t size(Args... args) const {
         auto target = Reference::size(args...);
         auto value = t.size(args...);
         if (target != value) {
@@ -61,7 +61,7 @@ public:
     }
 
     template<typename ...Args>
-    int select(Args... args) const {
+    int64_t select(Args... args) const {
         auto target = Reference::select(args...);
         auto value = t.select(args...);
         if (target != value) {
@@ -72,7 +72,7 @@ public:
     }
 
     template<typename ...Args>
-    int operator[](Args... args) const {
+    int64_t operator[](Args... args) const {
         auto target = Reference::operator[](args...);
         auto value = t.operator[](args...);
         if (target != value) {
@@ -96,17 +96,17 @@ class VectorWithRankSupport : public vector<T> {
 public:
     using vector<T>::vector;
     // rank [0..position]
-    int rank(T symbol,int position) const {
-        int result = 0;
+    int64_t rank(T symbol,int position) const {
+        int64_t result = 0;
         for (size_t i = 0; i <= position; ++i) {
             result += vector<T>::operator[](i) == symbol;
         }
         return result;
     }
 
-    int select(T symbol,int rank) const {
-        int crank = 1;
-        int i = 0;
+    int64_t select(T symbol,int64_t rank) const {
+        int64_t crank = 1;
+        int64_t i = 0;
         for(auto& c : *this) {
             if (c == symbol) {
                 if (crank == rank) {
@@ -121,11 +121,11 @@ public:
         return INT_MAX;
     }
 
-    char get(int position) const {
+    char get(int64_t position) const {
         return this->at(position);
     }
 
-    void insert(int position,T symbol) {
+    void insert(int64_t position,T symbol) {
         assert(position <= this->size());
         vector<T>::insert(this->begin() + position, symbol);
     }
@@ -137,14 +137,14 @@ class DynamicRoutingTableCore {
 public:
     DynamicRoutingTableCore(const DBGSuccinct& graph) : graph(graph) {}
 
-//    int select(node_index node, int occurrence, char symbol) const {
+//    int64_t select(node_index node, int64_t occurrence, char symbol) const {
 //    }
 
     // rank [0..position)
-    int rank(node_index node, int position, char symbol) const {
-        int encoded = graph.encode(symbol);
+    int64_t rank(node_index node, int64_t position, char symbol) const {
+        int64_t encoded = graph.encode(symbol);
         auto &node_entry = routing_table.at(node);
-        int result = 0;
+        int64_t result = 0;
         assert(position <= node_entry.size());
         if (position) {
             result = node_entry.rank(encoded, position - 1);
@@ -152,12 +152,12 @@ public:
         return result;
     }
 public:
-    int select(node_index node,int rank,char symbol) const {
-        int encoded = graph.encode(symbol);
+    int64_t select(node_index node,int64_t rank,char symbol) const {
+        int64_t encoded = graph.encode(symbol);
         auto &node_entry = routing_table.at(node);
         return node_entry.select(encoded,rank);
     }
-    char get(node_index node, int position) const {
+    char get(node_index node, int64_t position) const {
         auto &node_entry = routing_table.at(node);
         return graph.decode(node_entry[position]);
     }
@@ -165,7 +165,7 @@ public:
     string print_content(node_index node) const {
         stringstream out;
         auto table_size = size(node);
-        for (int i=0;i<table_size;i++) {
+        for (int64_t i=0;i<table_size;i++) {
             out << get(node, i);
         }
         out << endl;
@@ -173,26 +173,26 @@ public:
         return out.str();
     }
 
-    char traversed_base(node_index node, int position) const {
+    char traversed_base(node_index node, int64_t position) const {
         assert(position >= 0 && "traversing o");
         return get(node,position);
     }
 
-    int new_relative_position(node_index node, int position) const {
+    int64_t new_relative_position(node_index node, int64_t position) const {
         auto base = get(node,position);
         auto base_rank = rank(node,position,base);
         return base_rank;
     }
 
-    int size(node_index node) const {
+    int64_t size(node_index node) const {
         if (routing_table.count(node))
             return routing_table.at(node).size();
         else
             return 0;
     }
 
-    void insert(node_index node, int position, char symbol) {
-        int encoded = graph.encode(symbol);
+    void insert(node_index node, int64_t position, char symbol) {
+        int64_t encoded = graph.encode(symbol);
         assert(position <= size(node));
         routing_table[node].insert(position,encoded);
     }
